requirements.txt
Django==1.9.6
libfaketime==0.4.1
python-dateutil==2.4.1
